Move common/odp to net/
- "common" is used by all packages, incl. "net".
- "net" package is more appropriate place for odp helpers.
- Removes the need for "common/net".
