+++ net/if.go
@@ -82,3 +82,11 @@ func findInterface(ifaceName string) (iface *net.Interface, err error) {
 	}
 	return
 }
+
+func SetMTU(ifaceName string, mtu int) error {
+	link, err := netlink.LinkByName(ifaceName)
+	if err != nil {
+		return fmt.Errorf("cannot find interface %s: %s", ifaceName, err)
+	}
+	return netlink.LinkSetMTU(link, mtu)
+}
